fix: labelSlide virou Float (2º build da v0.19.4) — o CI cravou 3 erros de UM erro só: translationX é Float e o dp() devolve Int, então o '-labelSlide' no if/else com '0f' dobrou o tipo (Comparable&Number) e explodiu em ExpandingNavBar.kt:183 (Int but Float expected), :229 (Unresolved reference no '-' de tipo dobrado) e :247 (Comparable&Number but Float expected) — toFloat() no único lugar onde o slide nasce e os 3 caem juntos. Lógica zero mudada

+++ app/src/main/java/com/tunegrab/app/ui/ExpandingNavBar.kt
@@ -72,8 +72,10 @@ class ExpandingNavBar @JvmOverloads constructor(
 
     /** Deslize lateral do nome (v0.19.4): o texto SÓ anda pra frente/trás na
      * horizontal — nunca nasce de cima pra baixo (o singleLine do label mata
-     * a quebra de linha que fazia o texto "subir" durante a abertura). */
-    private val labelSlide = dp(10)
+     * a quebra de linha que fazia o texto "subir" durante a abertura).
+     * Float: translationX é Float e o dp() é Int — misturar os dois no
+     * if/else dobra o tipo (Comparable&Number) e o compilador crava erro. */
+    private val labelSlide = dp(10).toFloat()
 
     init {
         orientation = HORIZONTAL
